Adds client create endpoint
We want for trainers to be able to create new clients from the web UI.
We'll introduce a new 'POST /clients' endpoint to support this use case.

Here, we add the client model and database table, get a serializer in
place, and hook up a clients controller with the new create action.
Correctness of the new endpoint is tested with an end to end request
spec.