Adds create workout endpoint
Clients can have workouts that we want to be tracking in this
application, so we want to introduce the ability to create a new workout
belonging to a client.

Since workouts must always be created for a given client, we want to do
this as a nested resource.

Here, we go ahead and introduce the model, rig up the associations and
introduce the new workouts controller and create action, along with a
route to get this endpoint going.  Behavior is tested at the request
level.