Adds workout items to workouts
A workout is composed of many individual exercises that a client must
Workouts should be able to track the individual workout items on the
workout.  Let's go ahead and introduce workout items, which will just
have a title for now.

Here, we begin by setting up the new database table and the model on top
of it.  A new service is then added to handle the creation of the
workout items for a workout when it is created, and the service is
invoked from the controller.

This service exposes errors for the workout and workout item, the
merging of which requires a new error message merge mechanism that is
introduced as a refinement over ActiveModel::Error.

Behavior of this new workout item logic is tested at the request
spec level.