Adds result tracking for clients
Clients need to be able to track their results as they progress through
a workout.  To facilitate this, we're going to add a results property to
workout items, so clients can log their progression through a workout
there.

Here, we introduce a new workout item update endpoint which will allow
the UI to add results to the individual workout items on a workout.

This is done with the introduction of a results column on workout items,
a new workout items endpoint with some faked header-based client
authentication, and a namespaced route for the endpoint under a new
/clients namespace.  Behavior is tested at the request level.